backend-django/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==26.1.0
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron-descriptor==1.4.5
dataclasses-json==0.6.7
distro==1.9.0
Django==5.2.12
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-timezone-field==7.2.1
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
greenlet==3.3.2
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.0
kombu==5.6.2
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.20
langchain-groq==1.1.2
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.3
openai==2.29.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.3.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
sqlparse==0.5.5
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
vine==5.1.0
wcwidth==0.6.0
xxhash==3.6.0
yfinance==1.2.0
yarl==1.23.0
zstandard==0.25.0
